+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2-beta
+
+* Updated structure of package to use Builder
+* Updated example project to use the package and show a random avatar
+
 ## 0.0.1-beta
 
 * Versioned to beta and published on pub.dev